requirements.txt
argparse==1.4.0
humanize==4.7.0
imageio==2.31.1
matplotlib==3.9.1
nibabel==5.1.0
numpy==1.26.3
openpyxl==3.1.3
pandas==2.0.3
pillow==10.4.0
PyYAML==6.0.1
rich==13.5.1
# sam2 @ git+https://github.com/facebookresearch/sam2.git@2b90b9f
scikit-image==0.21.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
# Uncomment the following requirement after installing them so that PyCharm
# doesn't complain about a missing package
# torch==2.3.1
# segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509f
tqdm==4.66.4
